JGHG

+++ dbname.products.sql
@@ -9,9 +9,9 @@ CREATE TABLE IF NOT EXISTS dbname.products
     invoice_number VARCHAR(10),
     stock_code VARCHAR(20),
     detail VARCHAR(1000),
-    quantity INT,
+    quantity BIGINT,
     invoice_date TIMESTAMP,
     unit_price NUMERIC(10,5),
-    customer_id INT,
+    customer_id BIGINT,
     country VARCHAR(20) 
 );
